Refactor: Remove logic from Address Controller and secrets from application.properties

+++ src/main/java/com/nineleaps/leaps/service/implementation/AddressServiceImpl.java
@@ -1,13 +1,16 @@
 package com.nineleaps.leaps.service.implementation;
 
 import com.nineleaps.leaps.dto.AddressDto;
+import com.nineleaps.leaps.exceptions.AddressOwnershipException;
 import com.nineleaps.leaps.model.Address;
 import com.nineleaps.leaps.model.User;
 import com.nineleaps.leaps.repository.AddressRepository;
 import com.nineleaps.leaps.service.AddressServiceInterface;
+import com.nineleaps.leaps.utils.Helper;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
@@ -17,10 +20,12 @@ import java.util.Optional;
 @Transactional
 public class AddressServiceImpl implements AddressServiceInterface {
     private final AddressRepository addressRepository;
+    private final Helper helper;
 
     // Save a new address for the user
     @Override
-    public void saveAddress(AddressDto addressDto, User user) {
+    public void saveAddress(AddressDto addressDto, HttpServletRequest request) {
+        User user = helper.getUser(request);
         if (addressDto.isDefaultAddress()) {
             // If the new address is set as the default, remove the default flag from other addresses
             List<Address> addresses = addressRepository.findAllByUser(user);
@@ -35,7 +40,9 @@ public class AddressServiceImpl implements AddressServiceInterface {
 
     // List all addresses for the user
     @Override
-    public List<Address> listAddress(User user) {
+    public List<Address> listAddress(HttpServletRequest request) {
+        // JWT : Extracting user info from token
+        User user = helper.getUser(request);
         return addressRepository.findAllByUser(user);
     }
 
@@ -47,7 +54,9 @@ public class AddressServiceImpl implements AddressServiceInterface {
 
     // Read an address by its ID for a specific user
     @Override
-    public Address readAddress(User user, Long addressId) {
+    public Address readAddress(HttpServletRequest request, Long addressId) {
+        // JWT : Extracting user info from token
+        User user = helper.getUser(request);
         List<Address> body = addressRepository.findAllByUser(user);
         for (Address address : body) {
             if (addressId.equals(address.getId())) {
@@ -59,15 +68,24 @@ public class AddressServiceImpl implements AddressServiceInterface {
 
     // Update an existing address
     @Override
-    public void updateAddress(AddressDto addressDto, Long addressId, User user) {
+    public void updateAddress(AddressDto addressDto, Long addressId, HttpServletRequest request) {
+        Address checkAddress = readAddress(request, addressId);
+        if (checkAddress == null)
+            throw new AddressOwnershipException("Address does not belong to current user");
+
         addressDto.setId(addressId);
-        Address address = new Address(addressDto, user);
+        Address address = new Address(addressDto, helper.getUser(request));
         addressRepository.save(address);
     }
 
     // Delete an address by its ID
     @Override
-    public void deleteAddress(Long addressId) {
+    public void deleteAddress(HttpServletRequest request, Long addressId) {
+        // Guard Statement : Check if address belong to given user
+        Address checkAddress = readAddress(request, addressId);
+        if (checkAddress == null) {
+            throw new AddressOwnershipException("Address do not belong to the current user");
+        }
         addressRepository.deleteById(addressId);
     }
 }
